refactor: move dev APIs
